✨ feat: Add iPad image to cover page
Add the iPad image to the cover page of the website. The image is located at `_assets/images/ipad.png` and it is displayed using the Markdown syntax `![iPad](./_assets/images/ipad.png)`.

This commit also includes a corresponding change in the English version of the cover page, where the image path is adjusted to `../_assets/images/ipad.png`.

This addition enhances the visual appeal of the cover page and reinforces the message to return the lost iPad for rewards.